arreglar los directorios de los usuarios y cuando el token caduca desloguear al usuario
