+++ src/components/CabeceraCodecat.js
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 import { NavLink } from "react-router-dom";
-import { AuthContext } from "../../context/AuthContext";
+import { AuthContext } from "../context/AuthContext";
 
 export const CabeceraCodecat = () => {
   const { desloguearUsuario } = useContext(AuthContext);
