--- src/components/Loading/Loading.css
@@ -1,80 +1,0 @@
-:root {
-  --azul: rgb(41, 131, 233);
-}
-body {
-  height: 100vh;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen,
-    Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
-}
-
-.puntos {
-  position: relative;
-  left: 20px;
-  width: 7px;
-  height: 7px;
-  border-radius: 5px;
-  background-color: var(--azul);
-  color: var(--azul);
-  animation: puntosVariantes 0.8s infinite linear alternate;
-  animation-delay: 0.4s;
-}
-
-.puntos::before,
-.puntos::after {
-  content: "";
-  display: inline-block;
-  position: absolute;
-  top: 0;
-}
-
-.puntos::before {
-  left: -15px;
-  width: 7px;
-  height: 7px;
-  border-radius: 5px;
-  background-color: var(--azul);
-  color: var(--azul);
-  animation: puntosVariantes 0.8s infinite alternate;
-  animation-delay: 0s;
-}
-
-.puntos::after {
-  left: 15px;
-  width: 7px;
-  height: 7px;
-  border-radius: 5px;
-  background-color: var(--azul);
-  color: var(--azul);
-  animation: puntosVariantes 0.8s infinite alternate;
-  animation-delay: 0.8s;
-}
-
-@keyframes puntosVariantes {
-  0% {
-    background-color: var(--azul);
-  }
-  50%,
-  100% {
-    background-color: #ebe6ff;
-  }
-}
-.loading {
-  margin: 0 auto;
-  font-size: 20px;
-  font-weight: 700;
-  color: var(--azul);
-  display: flex;
-  flex-direction: row;
-  align-items: flex-end;
-}
-.contenedor-loading {
-  display: flex;
-  z-index: 2000;
-  justify-content: center;
-  align-items: center;
-  background-color: rgba(204, 202, 202, 0.61);
-  position: fixed;
-  width: 100vw;
-  top: 0;
-  bottom: 0;
-}
