+++ src/components/ejercicios/HeaderEjercicio.js
@@ -1,42 +1,32 @@
 import Button from "react-bootstrap/Button";
 import { useCallback, useContext, useState } from "react";
-import { AuthContext } from "../context/AuthContext";
-import "./HeaderEjercicio/HeaderEjercicio.css";
+import { AuthContext } from "../../context/AuthContext";
 import { Col } from "react-bootstrap";
-import { GeneralContext } from "../context/GeneralContext";
-import { EjerciciosContext } from "../context/EjerciciosContext";
-export const ListaSinOrdenar = (props) => {
+import { GeneralContext } from "../../context/GeneralContext";
+import { EjerciciosContext } from "../../context/EjerciciosContext";
+export const HeaderEjercicio = (props) => {
   const { datosEjercicio, idTrabajo } = props;
+  const { conseguirExperiencia } = useContext(EjerciciosContext);
   const { token } = useContext(AuthContext);
   const { urlApi } = useContext(GeneralContext);
-  const { conseguirExperiencia } = useContext(EjerciciosContext);
-
   const [input1, setInput1] = useState("");
-  const [input2, setInput2] = useState("");
-  const [input3, setInput3] = useState("");
-
   const [acierto, setAcierto] = useState(false);
   const [texto, setTexto] = useState("");
 
   const comprobarEjercicio = useCallback(async () => {
-    if (
-      input1 === datosEjercicio.objetivos[0] &&
-      input2 === datosEjercicio.objetivos[1] &&
-      input3 === datosEjercicio.objetivos[2]
-    ) {
+    if (input1 === datosEjercicio.objetivos[0]) {
       setAcierto(true);
-      const response = await fetch(
+      await fetch(
         `${urlApi}historial/anyadir-tarea/${idTrabajo}/${datosEjercicio._id}`,
         {
           method: "PUT",
           headers: { Authorization: "Bearer " + token },
         }
       );
-      await response.json();
       await conseguirExperiencia(datosEjercicio.recompensa.experiencia);
       setTexto("");
     } else {
-      setTexto("Esta lista no es para fliparlo");
+      setTexto("Este header no es tan chulo!!");
       setAcierto(false);
     }
   }, [
@@ -46,8 +36,6 @@ export const ListaSinOrdenar = (props) => {
     datosEjercicio.recompensa.experiencia,
     idTrabajo,
     input1,
-    input2,
-    input3,
     token,
     urlApi,
   ]);
@@ -62,30 +50,10 @@ export const ListaSinOrdenar = (props) => {
         <input type="text" onChange={(e) => setInput1(e.target.value)} />
         <span>{" >"}</span>
       </Col>
-
-      <Col xs="8" className="input-ejercicio">
-        <span>{"< "}</span>
-        <input type="text" onChange={(e) => setInput2(e.target.value)} />
-        <span>{" >"}</span>
-      </Col>
-      <Col xs="12" className="text-right">
-        Primer elemento
-      </Col>
-      <Col xs="8" className="input-ejercicio">
-        <span>{"</ "}</span>
-        <input type="text" value={input2} readOnly />
-        <span>{" >"}</span>
-      </Col>
-      <Col xs="8" className="input-ejercicio">
-        <span>{"< "}</span>
-        <input type="text" onChange={(e) => setInput3(e.target.value)} />
-        <span>{" >"}</span>
-      </Col>
-      <Col xs="12">Segundo elemento</Col>
-      <Col xs="8" className="input-ejercicio  justify-self-end">
-        <span>{"</ "}</span>
-        <input type="text" value={input3} readOnly />
-        <span>{" >"}</span>
+      <Col xs="12">
+        <span>{"<h1>"}</span>
+        <span>Esta es una cabecera</span>
+        <span>{" </h1>"}</span>
       </Col>
       <Col xs="12" className="input-ejercicio ultimo-input">
         <span>{"</ "} </span>
